i18n optimization, use capitalized key as default english translation if not found.
