+++ src/components/common/404.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import { Button } from 'antd'
-import t from './i18n'
+import t from '../../i18n'
 
 import './404.less'
