--- src/components/common/i18n.js
@@ -1,301 +1,0 @@
-// string format
-function format (txt) {
-  var args = [].slice.call(arguments, 1)
-  return txt.replace(/{(\d+)}/g, function (m, i) {
-    return args[i]
-  })
-}
-
-function t (code, args) {
-  var txt = msgs[lang][code] || code
-  return args ? format(txt, args) : txt
-}
-
-export default t
-
-export const lang = function () {
-  var lang = localStorage.getItem('lang') || navigator.language
-  localStorage.setItem('lang', lang)
-  return lang.substr(0, 2)
-}()
-
-const msgs = {
-  'en': {
-    '404': 'Page is not found',
-    'error.info': 'Error Info:',
-    'back.home': 'Back to home',
-    'exit': 'Exit',
-    'submit': 'Submit',
-    'confirm': 'Confirm',
-    'cancel': 'Cancel',
-    'close': 'Close',
-    'add': 'Add',
-    'refresh': 'Refresh',
-    'update': 'Update',
-    'delete': 'Delete',
-    'enable': 'Enable',
-    'disable': 'Disable',
-    'finished': 'Finished',
-    'ajax.failed': 'Request failed',
-    'ajax.success': 'Request success',
-    'operate.success': 'Operate success',
-    'on': 'ON',
-    'off': 'OFF',
-    'input': 'Please Input',
-    'trigger': 'Trigger',
-    'monitor': 'Monitor',
-    'pause': 'Pause',
-    'resume': 'Resume',
-    'stop': 'Stop',
-    'terminate': 'Terminate',
-    'schedule': 'Schedule',
-    'not.found': 'Not found records',
-    'apps.mgr': 'App Management',
-    'apps.list': 'App List',
-    'apps.edit': 'Edit App',
-    'apps.delete': 'Delete App',
-    'apps.select': 'Select App',
-    'apps.name': 'App Name',
-    'apps.desc': 'App Description',
-    'apps.key': 'App Secret Key',
-    'apps.inherit': 'Inherit Configs From App',
-    'apps.delete.confirm': 'Sure to delete the app【{0}】and delete all configurations of it?',
-    'jobs.mgr': 'Job Management',
-    'jobs.configs': 'Job Config',
-    'jobs.controls': 'Job Control',
-    'jobs.instances': 'Job History',
-    'jobs.edit': 'Edit Job',
-    'jobs.list': 'Job List',
-    'jobs.running': 'Job Monitor',
-    'jobs.count': 'Scheduling Job Count',
-    'jobs.cron': 'Cron Expression',
-    'jobs.class': 'Job Class',
-    'jobs.class.exapmle': 'e.g., com.xxx.myJob',
-    'jobs.desc': 'Job Description',
-    'jobs.params': 'Params',
-    'jobs.shard.count': 'Shard Count',
-    'jobs.shard.params': 'Shard Params',
-    'jobs.timeout': 'Timeout(Second)',
-    'jobs.shard.params.exapmle': 'e.g., 0=A;1=B;2=C',
-    'jobs.misfire': 'Misfire or not',
-    'jobs.max.shard.pull.count': 'Max Shard Pull Count',
-    'jobs.trigger': 'Trigger Job',
-    'jobs.pause': 'Pause Job',
-    'jobs.resume': 'Resume Job',
-    'jobs.disable': 'Disable Job',
-    'jobs.enable': 'Enable Job',
-    'jobs.delete': 'Delete Job',
-    'jobs.schedule': 'Schedule Job',
-    'jobs.stop': 'Stop Job',
-    'jobs.terminate': 'Terminate Job',
-    'jobs.del_next': 'Delete the next job',
-    'jobs.trigger.confirm': 'Sure to trigger the job?',
-    'jobs.pause.confirm': 'Sure to pause the job?',
-    'jobs.resume.confirm': 'Sure to resume the job?',
-    'jobs.disable.confirm': 'Sure to disable the job?',
-    'jobs.enable.confirm': 'Sure to enable the job?',
-    'jobs.delete.confirm': 'Sure to delete the job?',
-    'jobs.schedule.confirm': 'Sure to schedule the job?',
-    'jobs.stop.confirm': 'Sure to stop the job?',
-    'jobs.del_next.confirm': 'Sure to delete the next job?',
-    'jobs.terminate.confirm': 'Sure to terminate the current job instance?',
-    'jobs.fire.time.prev': 'Previous FireTime',
-    'jobs.fire.time': 'Current FireTime',
-    'jobs.fire.time.next': 'Next FireTime',
-    'jobs.scheduler': 'Schedule Server',
-    'jobs.instance.detail': 'Job Instance Detail',
-    'jobs.instance.running.info': 'Running Progress',
-    'jobs.instance.shards.info': 'Shards Detail',
-    'jobs.instance.finish.percent': 'Finish Percent',
-    'jobs.instance.total.shard.count': 'Total Shard Count',
-    'jobs.instance.wait.shard.count': 'Wait Shard Count',
-    'jobs.instance.running.shard.count': 'Running Shard Count',
-    'jobs.instance.success.shard.count': 'Success Shard Count',
-    'jobs.instance.failed.shard.count': 'Failed Shard Count',
-    'jobs.instance.shard.pull.client': 'Pull Client:',
-    'jobs.instance.shard.finish.client': 'Finish Client:',
-    'jobs.instance.shard.pull.count': 'Pull Count:',
-    'jobs.instance.shard.pull.time': 'Pull Time:',
-    'jobs.instance.shard.item': 'Shard Item',
-    'jobs.instance.shard.param': 'Shard Param',
-    'jobs.config.info': 'Config Info',
-    'jobs.basic.info': 'Basic Info',
-    'jobs.dependence': '[{0}]\'s Next Jobs',
-    'jobs.dependence.config': 'Dependence Config',
-    'jobs.next.ids': 'Next Job Ids, e.g., 123,456',
-    'jobs.assign': 'Assign【{0}】\'s clients',
-    'jobs.assigns': 'Assign Client',
-    'jobs.assign.inst': 'Instance Number',
-    'clusters.mgr': 'Cluster Management',
-    'clusters.servers': 'Server List',
-    'clusters.servers.jobs': 'Server[{0}]\'s jobs',
-    'clusters.clients': 'Client List',
-    'start.time': 'Start Time:',
-    'end.time': 'End Time:',
-    'table.start.time': 'Start Time',
-    'table.end.time': 'End Time',
-    'table.cost.time': 'Cost Time',
-    'table.trigger.type': 'Trigger Type',
-    'id': 'ID',
-    'name': 'Name',
-    'desc': 'Description',
-    'operation': 'Operation',
-    'value': 'Value',
-    'total': 'All {0} items',
-    'input.fullname': 'Pleanse Input Full Name',
-    'input.classname': 'Pleanse Input Job Class Full Name',
-    'clients.addr': 'Client Address',
-    'host': 'Host',
-    'status': 'Status',
-    'shutdown': 'Shutdown',
-    'clean.cache': 'Clean Cache',
-    'servers.shutdown': 'Shutdown Server',
-    'servers.shutdown.confirm': 'Sure to shutdown the server【{0}】?',
-    'servers.jobs.list': 'Job List',
-    'name.tip': '(A-Za-z0-9_)',
-    'field.format.error': 'The format isn\'t right: {0}',
-    'enable.or.not': 'Enable or not',
-    'ip': 'IP'
-  },
-
-  'zh': {
-    '404': '未找到该页面',
-    'error.info': '错误信息：',
-    'back.home': '返回首页',
-    'exit': '退出',
-    'submit': '提交',
-    'confirm': '确认',
-    'cancel': '取消',
-    'close': '关闭',
-    'add': '添加',
-    'refresh': '刷新',
-    'update': '更新',
-    'delete': '删除',
-    'input': '请输入',
-    'trigger': '触发',
-    'pause': '暂停',
-    'resume': '恢复',
-    'monitor': '监控',
-    'stop': '停止',
-    'schedule': '调度',
-    'terminate': '终止',
-    'enable': '启用',
-    'disable': '停用',
-    'finished': '完成',
-    'ajax.failed': '请求失败',
-    'ajax.success': '请求成功',
-    'operate.success': '操作成功',
-    'more': '更多',
-    'on': 'ON',
-    'off': 'OFF',
-    'not.found': '未找到记录',
-    'apps.mgr': '应用管理',
-    'apps.list': '应用列表',
-    'apps.edit': '编辑应用',
-    'apps.delete': '删除应用',
-    'apps.select': '请选择应用',
-    'apps.name': '应用名称',
-    'apps.desc': '应用描述',
-    'apps.key': '应用密钥',
-    'apps.delete.confirm': '确定要删除应用【{0}】，并删除其所有配置吗？',
-    'apps.inherit': '继承应用配置',
-    'jobs.mgr': '任务管理',
-    'jobs.configs': '任务配置',
-    'jobs.controls': '任务管控',
-    'jobs.instances': '任务历史',
-    'jobs.edit': '编辑任务',
-    'jobs.list': '任务列表',
-    'jobs.running': '任务监控',
-    'jobs.count': '调度任务数',
-    'jobs.cron': 'Cron表达式',
-    'jobs.class': '任务处理类',
-    'jobs.class.exapmle': '如com.xxx.myJob',
-    'jobs.desc': '任务描述',
-    'jobs.params': '自定义参数',
-    'jobs.shard.count': '分片数',
-    'jobs.shard.params': '分片参数',
-    'jobs.shard.params.exapmle': '如0=A;1=B;2=C',
-    'jobs.misfire': '支持Misfire',
-    'jobs.timeout': '超时设置(秒)',
-    'jobs.max.shard.pull.count': '分片最大拉取次数',
-    'jobs.trigger': '触发任务',
-    'jobs.pause': '暂停任务',
-    'jobs.resume': '恢复任务',
-    'jobs.disable': '停用任务',
-    'jobs.enable': '启用任务',
-    'jobs.delete': '删除任务',
-    'jobs.schedule': '调度任务',
-    'jobs.stop': '停止任务',
-    'jobs.del_next': '删除后置任务',
-    'jobs.terminate': '终止任务',
-    'jobs.trigger.confirm': '确定要触发一次该任务吗，操作后任务将执行一次？',
-    'jobs.pause.confirm': '确定要暂停执行该任务吗，暂停后任务将不被调度，可通过【恢复】操作重新调度？',
-    'jobs.resume.confirm': '确定要恢复执行该任务吗，恢复后任务将开始调度？',
-    'jobs.disable.confirm': '确定要停用该任务吗，停用后任务将不再被调度，可通过【启用】操作重新调度',
-    'jobs.enable.confirm': '确定要启用该任务吗，启用后任务将开始调度？',
-    'jobs.delete.confirm': '确定要删除该任务吗？',
-    'jobs.stop.confirm': '确定要删除停止该任务吗，停止后任务不再被调度，可通过【调度】操作重新调度？',
-    'jobs.schedule.confirm': '确定要开始调度该任务吗，之后任务将开始调度？',
-    'jobs.del_next.confirm': '确定要删除该后置任务？',
-    'jobs.terminate.confirm': '确定要终止当前任务实例吗？',
-    'jobs.fire.time.prev': '上次触发时间',
-    'jobs.fire.time': '本次触发时间',
-    'jobs.fire.time.next': '下次触发时间',
-    'jobs.scheduler': '调度服务器',
-    'jobs.instance.detail': '任务实例详情',
-    'jobs.config.info': '配置信息',
-    'jobs.basic.info': '基本信息',
-    'jobs.instance.running.info': '运行进度',
-    'jobs.instance.shards.info': '分片详情',
-    'jobs.instance.finish.percent': '完成百分比',
-    'jobs.instance.total.shard.count': '总分片数',
-    'jobs.instance.wait.shard.count': '待执行',
-    'jobs.instance.running.shard.count': '执行中',
-    'jobs.instance.success.shard.count': '成功',
-    'jobs.instance.failed.shard.count': '失败',
-    'jobs.instance.shard.item': '分片项',
-    'jobs.instance.shard.param': '分片参数',
-    'jobs.instance.shard.pull.client': '拉取客户端：',
-    'jobs.instance.shard.finish.client': '完成客户端：',
-    'jobs.instance.shard.pull.count': '拉取次数：',
-    'jobs.instance.shard.pull.time': '拉取时间：',
-    'jobs.dependence': '【{0}】的后置任务',
-    'jobs.dependence.config': '依赖配置',
-    'jobs.next.ids': '后置任务ID, 如123',
-    'jobs.assign': '指定【{0}】的客户端',
-    'jobs.assigns': '指定客户端',
-    'jobs.assign.inst': '实例数',
-    'clusters.mgr': '集群管理',
-    'clusters.servers': '服务器列表',
-    'clusters.servers.jobs': '服务器【{0}】调度的任务',
-    'clusters.clients': '客户端列表',
-    'start.time': '开始时间：',
-    'end.time': '结束时间：',
-    'table.start.time': '开始执行时间',
-    'table.end.time': '结束执行时间',
-    'table.cost.time': '总耗时',
-    'table.trigger.type': '触发方式',
-    'id': '编号',
-    'name': '名称',
-    'desc': '描述',
-    'operation': '操作',
-    'value': '值',
-    'total': '共 {0} 条',
-    'input.fullname': '请输入全名称',
-    'input.classname': '请输入Job全类名',
-    'clients.addr': '客户端地址',
-    'host': '主机',
-    'status': '状态',
-    'shutdown': '关闭',
-    'servers.jobs.list': '查看任务列表',
-    'servers.clean.cache': '清除服务器缓存',
-    'servers.clean.cache.confirm': '确定要清除服务器【{0}】的缓存吗？',
-    'servers.shutdown': '关闭服务器',
-    'servers.shutdown.confirm': '确定要关闭服务器【{0}】吗？',
-    'name.tip': '(字母、数字或下划线)',
-    'field.format.error': '输入格式不正确: {0}',
-    'enable.or.not': '是否启用',
-    'ip': 'IP'
-  }
-}
-
