+++ src/components/layouts/layout-header.less
@@ -5,7 +5,7 @@
   background: #404040;
 
   .logo {
-    background: url("../assets/img/logo.png");
+    background: url("../../assets/img/logo.png");
     width: 100px;
     height: 64px;
     float: left;
