+++ src/i18n/en.js
@@ -1,0 +1,119 @@
+export default {
+
+  '404': 'Page is not found',
+
+  'ajax.failed': 'Request failed',
+  'ajax.success': 'Request success',
+
+  'apps.mgr': 'App Management',
+  'apps.list': 'App List',
+  'apps.edit': 'Edit App',
+  'apps.delete': 'Delete App',
+  'apps.select': 'Select App',
+  'apps.name': 'App Name',
+  'apps.desc': 'App Description',
+  'apps.key': 'App Secret Key',
+  'apps.inherit': 'Inherit Configs From App',
+  'apps.delete.confirm': 'Sure to delete the app【{0}】and delete all configurations of it?',
+
+  'clients.addr': 'Client Address',
+  'clusters.mgr': 'Cluster Management',
+  'clusters.servers': 'Server List',
+  'clusters.servers.jobs': 'Server[{0}]\'s jobs',
+  'clusters.clients': 'Client List',
+
+  'back.home': 'Back to home',
+
+  'desc': 'Description',
+
+  'error.info': 'Error Info:',
+  'field.format.error': 'The format isn\'t right: {0}',
+
+  'id': 'ID',
+  'ip': 'IP',
+  'input': 'Please Input',
+  'input.fullname': 'Please Input Full Name',
+  'input.classname': 'Please Input Job Class Full Name',
+
+  'name.tip': '(A-Za-z0-9_)',
+  'not.found': 'Not found records',
+
+  'jobs.mgr': 'Job Management',
+  'jobs.configs': 'Job Config',
+  'jobs.controls': 'Job Control',
+  'jobs.instances': 'Job History',
+  'jobs.edit': 'Edit Job',
+  'jobs.running': 'Job Monitor',
+  'jobs.count': 'Scheduling Job Count',
+  'jobs.cron': 'Cron Expression',
+  'jobs.class.exapmle': 'e.g., com.xxx.myJob',
+  'jobs.desc': 'Job Description',
+  'jobs.params': 'Params',
+  'jobs.shard.count': 'Shard Count',
+  'jobs.shard.params': 'Shard Params',
+  'jobs.timeout': 'Timeout(Second)',
+  'jobs.shard.params.exapmle': 'e.g., 0=A;1=B;2=C',
+  'jobs.misfire': 'Misfire or not',
+  'jobs.max.shard.pull.count': 'Max Shard Pull Count',
+  'jobs.trigger': 'Trigger Job',
+  'jobs.pause': 'Pause Job',
+  'jobs.resume': 'Resume Job',
+  'jobs.disable': 'Disable Job',
+  'jobs.enable': 'Enable Job',
+  'jobs.delete': 'Delete Job',
+  'jobs.schedule': 'Schedule Job',
+  'jobs.stop': 'Stop Job',
+  'jobs.terminate': 'Terminate Job',
+  'jobs.del_next': 'Delete the next job',
+  'jobs.trigger.confirm': 'Sure to trigger the job?',
+  'jobs.pause.confirm': 'Sure to pause the job?',
+  'jobs.resume.confirm': 'Sure to resume the job?',
+  'jobs.disable.confirm': 'Sure to disable the job?',
+  'jobs.enable.confirm': 'Sure to enable the job?',
+  'jobs.delete.confirm': 'Sure to delete the job?',
+  'jobs.schedule.confirm': 'Sure to schedule the job?',
+  'jobs.stop.confirm': 'Sure to stop the job?',
+  'jobs.del_next.confirm': 'Sure to delete the next job?',
+  'jobs.terminate.confirm': 'Sure to terminate the current job instance?',
+  'jobs.fire.time.prev': 'Previous FireTime',
+  'jobs.fire.time': 'Current FireTime',
+  'jobs.fire.time.next': 'Next FireTime',
+  'jobs.scheduler': 'Schedule Server',
+  'jobs.instance.detail': 'Job Instance Detail',
+  'jobs.instance.running.info': 'Running Progress',
+  'jobs.instance.shards.info': 'Shards Detail',
+  'jobs.instance.finish.percent': 'Finish Percent',
+  'jobs.instance.total.shard.count': 'Total Shard Count',
+  'jobs.instance.wait.shard.count': 'Wait Shard Count',
+  'jobs.instance.running.shard.count': 'Running Shard Count',
+  'jobs.instance.success.shard.count': 'Success Shard Count',
+  'jobs.instance.failed.shard.count': 'Failed Shard Count',
+  'jobs.instance.shard.pull.client': 'Pull Client:',
+  'jobs.instance.shard.finish.client': 'Finish Client:',
+  'jobs.instance.shard.pull.count': 'Pull Count:',
+  'jobs.instance.shard.pull.time': 'Pull Time:',
+  'jobs.instance.shard.item': 'Shard Item',
+  'jobs.instance.shard.param': 'Shard Param',
+  'jobs.config.info': 'Config Info',
+  'jobs.basic.info': 'Basic Info',
+  'jobs.dependence': '[{0}]\'s Next Jobs',
+  'jobs.dependence.config': 'Dependence Config',
+  'jobs.next.ids': 'Next Job Ids, e.g., 123,456',
+  'jobs.assign': 'Assign【{0}】\'s clients',
+  'jobs.assigns': 'Assign Client',
+  'jobs.assign.inst': 'Instance Number',
+
+  'operate.success': 'Operate success',
+  'on': 'ON',
+  'off': 'OFF',
+
+  'servers.shutdown': 'Shutdown Server',
+  'servers.shutdown.confirm': 'Sure to shutdown the server【{0}】?',
+  'servers.jobs.list': 'Job List',
+
+  'table.start.time': 'Start Time',
+  'table.end.time': 'End Time',
+  'table.cost.time': 'Cost Time',
+  'table.trigger.type': 'Trigger Type',
+  'total': 'All {0} items',
+}
