+++ src/i18n/zh.js
@@ -1,0 +1,153 @@
+export default {
+  '404': '未找到该页面',
+  'add': '添加',
+  'ajax.failed': '请求失败',
+  'ajax.success': '请求成功',
+  'apps.mgr': '应用管理',
+  'apps.list': '应用列表',
+  'apps.edit': '编辑应用',
+  'apps.delete': '删除应用',
+  'apps.select': '请选择应用',
+  'apps.name': '应用名称',
+  'apps.desc': '应用描述',
+  'apps.key': '应用密钥',
+  'apps.delete.confirm': '确定要删除应用【{0}】，并删除其所有配置吗？',
+  'apps.inherit': '继承应用配置',
+
+  'back.home': '返回首页',
+
+  'confirm': '确认',
+  'cancel': '取消',
+  'close': '关闭',
+  'clients.addr': '客户端地址',
+  'clusters.mgr': '集群管理',
+  'clusters.servers': '服务器列表',
+  'clusters.servers.jobs': '服务器【{0}】调度的任务',
+  'clusters.clients': '客户端列表',
+
+  'desc': '描述',
+  'delete': '删除',
+  'disable': '停用',
+  'enable': '启用',
+  'enable.or.not': '是否启用',
+  'end.time': '结束时间：',
+  'error.info': '错误信息：',
+  'exit': '退出',
+
+  'field.format.error': '输入格式不正确: {0}',
+  'finished': '完成',
+
+  'host': '主机',
+
+  'id': '编号',
+  'ip': 'IP',
+  'input': '请输入',
+  'input.fullname': '请输入全名称',
+  'input.classname': '请输入Job全类名',
+
+  'jobs.mgr': '任务管理',
+  'jobs.configs': '任务配置',
+  'jobs.controls': '任务管控',
+  'jobs.instances': '任务历史',
+  'jobs.edit': '编辑任务',
+  'jobs.list': '任务列表',
+  'jobs.running': '任务监控',
+  'jobs.count': '调度任务数',
+  'jobs.cron': 'Cron表达式',
+  'jobs.class': '任务处理类',
+  'jobs.class.exapmle': '如com.xxx.myJob',
+  'jobs.desc': '任务描述',
+  'jobs.params': '自定义参数',
+  'jobs.shard.count': '分片数',
+  'jobs.shard.params': '分片参数',
+  'jobs.shard.params.exapmle': '如0=A;1=B;2=C',
+  'jobs.misfire': '支持Misfire',
+  'jobs.timeout': '超时设置(秒)',
+  'jobs.max.shard.pull.count': '分片最大拉取次数',
+  'jobs.trigger': '触发任务',
+  'jobs.pause': '暂停任务',
+  'jobs.resume': '恢复任务',
+  'jobs.disable': '停用任务',
+  'jobs.enable': '启用任务',
+  'jobs.delete': '删除任务',
+  'jobs.schedule': '调度任务',
+  'jobs.stop': '停止任务',
+  'jobs.del_next': '删除后置任务',
+  'jobs.terminate': '终止任务',
+  'jobs.trigger.confirm': '确定要触发一次该任务吗，操作后任务将执行一次？',
+  'jobs.pause.confirm': '确定要暂停执行该任务吗，暂停后任务将不被调度，可通过【恢复】操作重新调度？',
+  'jobs.resume.confirm': '确定要恢复执行该任务吗，恢复后任务将开始调度？',
+  'jobs.disable.confirm': '确定要停用该任务吗，停用后任务将不再被调度，可通过【启用】操作重新调度',
+  'jobs.enable.confirm': '确定要启用该任务吗，启用后任务将开始调度？',
+  'jobs.delete.confirm': '确定要删除该任务吗？',
+  'jobs.stop.confirm': '确定要删除停止该任务吗，停止后任务不再被调度，可通过【调度】操作重新调度？',
+  'jobs.schedule.confirm': '确定要开始调度该任务吗，之后任务将开始调度？',
+  'jobs.del_next.confirm': '确定要删除该后置任务？',
+  'jobs.terminate.confirm': '确定要终止当前任务实例吗？',
+  'jobs.fire.time.prev': '上次触发时间',
+  'jobs.fire.time': '本次触发时间',
+  'jobs.fire.time.next': '下次触发时间',
+  'jobs.scheduler': '调度服务器',
+  'jobs.instance.detail': '任务实例详情',
+  'jobs.config.info': '配置信息',
+  'jobs.basic.info': '基本信息',
+  'jobs.instance.running.info': '运行进度',
+  'jobs.instance.shards.info': '分片详情',
+  'jobs.instance.finish.percent': '完成百分比',
+  'jobs.instance.total.shard.count': '总分片数',
+  'jobs.instance.wait.shard.count': '待执行',
+  'jobs.instance.running.shard.count': '执行中',
+  'jobs.instance.success.shard.count': '成功',
+  'jobs.instance.failed.shard.count': '失败',
+  'jobs.instance.shard.item': '分片项',
+  'jobs.instance.shard.param': '分片参数',
+  'jobs.instance.shard.pull.client': '拉取客户端：',
+  'jobs.instance.shard.finish.client': '完成客户端：',
+  'jobs.instance.shard.pull.count': '拉取次数：',
+  'jobs.instance.shard.pull.time': '拉取时间：',
+  'jobs.dependence': '【{0}】的后置任务',
+  'jobs.dependence.config': '依赖配置',
+  'jobs.next.ids': '后置任务ID, 如123',
+  'jobs.assign': '指定【{0}】的客户端',
+  'jobs.assigns': '指定客户端',
+  'jobs.assign.inst': '实例数',
+
+  'pause': '暂停',
+  'monitor': '监控',
+  'more': '更多',
+  'not.found': '未找到记录',
+  'name.tip': '(字母、数字或下划线)',
+  'name': '名称',
+
+  'operate.success': '操作成功',
+  'operation': '操作',
+  'on': 'ON',
+  'off': 'OFF',
+
+  'resume': '恢复',
+  'refresh': '刷新',
+
+  'stop': '停止',
+  'schedule': '调度',
+  'submit': '提交',
+  'start.time': '开始时间：',
+  'switch': '切换',
+  'status': '状态',
+  'servers.jobs.list': '查看任务列表',
+  'servers.clean.cache': '清除服务器缓存',
+  'servers.clean.cache.confirm': '确定要清除服务器【{0}】的缓存吗？',
+  'servers.shutdown': '关闭服务器',
+  'servers.shutdown.confirm': '确定要关闭服务器【{0}】吗？',
+  'shutdown': '关闭',
+
+  'terminate': '终止',
+  'table.start.time': '开始执行时间',
+  'table.end.time': '结束执行时间',
+  'table.cost.time': '总耗时',
+  'table.trigger.type': '触发方式',
+  'total': '共 {0} 条',
+  'trigger': '触发',
+
+  'update': '更新',
+  'value': '值',
+}
